applied feedback from pull request
